requirements.txt
asgiref==3.3.1
Django==3.1.6
django-phonenumber-field==5.0.0
gunicorn==20.0.4
phonenumbers==8.12.17
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
